rename to PathStyleExtension
