+++ docs/api-reference/extensions/overview.md
@@ -11,6 +11,7 @@ This module contains the following extensions:
 - [BrushingExtension](/docs/api-reference/extensions/brushing-extension.md)
 - [DataFilterExtension](/docs/api-reference/extensions/data-filter-extension.md)
 - [Fp64Extension](/docs/api-reference/extensions/fp64-extension.md)
+- [PathStyleExtension](/docs/api-reference/extensions/path-style-extension.md)
 
 For instructions on authoring your own layer extensions, visit [developer guide](/docs/developer-guide/custom-layers/layer-extensions.md).
