+++ modules/extensions/src/index.js
@@ -1,4 +1,4 @@
 export {default as BrushingExtension} from './brushing/brushing';
 export {default as DataFilterExtension} from './data-filter/data-filter';
 export {default as Fp64Extension} from './fp64/fp64';
-export {default as PathExtension} from './path/path';
+export {default as PathStyleExtension} from './path-style/path-style';
