+++ modules/extensions/src/path-style/shaders.glsl.js
