+++ modules/layers/src/polygon-layer/polygon-layer.js
@@ -63,7 +63,7 @@ export default class PolygonLayer extends CompositeLayer {
     };
 
     if (this.props.getLineDashArray) {
-      log.removed('getLineDashArray', 'PathExtension')();
+      log.removed('getLineDashArray', 'PathStyleExtension')();
     }
   }
