+++ website/contents/pages.js
@@ -679,6 +679,10 @@ export const docPages = generatePath([
             content: getDocUrl('api-reference/extensions/overview.md')
           },
           {
+            name: 'BrushingExtension',
+            content: getDocUrl('api-reference/extensions/brushing-extension.md')
+          },
+          {
             name: 'DataFilterExtension',
             content: getDocUrl('api-reference/extensions/data-filter-extension.md')
           },
@@ -687,8 +691,8 @@ export const docPages = generatePath([
             content: getDocUrl('api-reference/extensions/fp64-extension.md')
           },
           {
-            name: 'BrushingExtension',
-            content: getDocUrl('api-reference/extensions/brushing-extension.md')
+            name: 'PathStyleExtension',
+            content: getDocUrl('api-reference/extensions/path-style-extension.md')
           }
         ]
       },
